+++ assets/js/event/rdv/calendar-index.js
@@ -1,7 +1,5 @@
 import CalendarManager from './CalendarManager'
-import SearchLocation from '../../utils/searchLocation'
 
 window.addEventListener('load', () => {
     new CalendarManager()
-    new SearchLocation('rdv_search_location')
 })