+++ assets/js/location-search.js
@@ -1,0 +1,3 @@
+import LocationSearcher from './utils/LocationSearcher'
+
+document.querySelectorAll('[data-location-search]').forEach(elt => new LocationSearcher(elt))