+++ src/Form/Evaluation/EvaluationGroupType.php
@@ -62,7 +62,7 @@ class EvaluationGroupType extends AbstractType
     {
         $resolver->setDefaults([
             'data_class' => EvaluationGroup::class,
-                'translation_domain' => 'evaluation',
+            'translation_domain' => 'evaluation',
         ]);
     }